set ambiwidth single

+++ init.lua
@@ -6,7 +6,7 @@ vim.g.loaded_netrw = 1
 vim.g.loaded_netrwPlugin = 1
 
 vim.o.number = true
-vim.o.ambiwidth = "double"
+vim.o.ambiwidth = "single"
 vim.o.tabstop = 2
 vim.o.shiftwidth = 2
 vim.o.softtabstop = 2
